Added serve command setting
Added a setting in the preferences window to store the command used when the serve button is clicked on, and set the serve button to use that command. Also added some code to the Travis CI config to install the schema for the program in order to prevent a related error.

Closes #16