Added screenshots page
Added a page with a few screenshots of the program. Also added a link to the main page to the top heading and some css to format that link.

Added a Screenshots section to the main page with a link to the screenshots page. Also added a new sub-directory to the images directory to store all of the screenshot images.

Closes #29